requirements.txt
flask==3.1.0
gunicorn==23.0.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
anthropic==0.52.0
requests==2.32.3
